+++ lib/elixir/lib/kernel/typespec.ex
@@ -583,7 +583,7 @@ defmodule Kernel.Typespec do
   # Handle access macro
   defp typespec({{:., meta, [Kernel, :access]}, meta1, [target, args]}, vars, caller) do
     access = {{:., meta, [Kernel, :access]}, meta1,
-              [target, args ++ [_: (quote hygiene: false, do: any)]]}
+              [target, args ++ [_: { :any, [], [] }]]}
     typespec(Macro.expand(access, caller), vars, caller)
   end